10.13.23 User is able to view multiple recipes.
Will start cleaning this up next. User will be able to view multiple recipes, but instead of seeing ingredients and instructions for various recipes all in one page, I will work on displaying the results recipe name + image. User will then be able to pick a recipe and they will be taken to a new page that has all the important recipe information.